requirements.txt
google-generativeai==0.8.3
requests==2.31.0
pandas==2.1.0
numpy==1.24.0
beautifulsoup4==4.12.2
feedparser==6.0.10 